+++ src/main/java/edith/Edith.java
@@ -1,3 +1,5 @@
+package edith;
+
 import java.util.Objects;
 import java.util.Scanner;
