+++ src/main/java/edith/Parser.java
@@ -1,4 +1,9 @@
-import javafx.beans.binding.ObjectExpression;
+package edith;
+
+import edith.exception.MissingDeadlineException;
+import edith.exception.MissingEventDurationException;
+import edith.exception.MissingTaskNameException;
+import edith.exception.MissingTaskNumberException;
 
 import java.util.Arrays;
 import java.util.List;
