+++ src/main/java/edith/exception/MissingDeadlineException.java
@@ -1,4 +1,4 @@
-package Edith;
+package edith.exception;
 
 public class MissingDeadlineException extends Exception {
     public MissingDeadlineException() {
