+++ src/main/java/edith/task/ToDoTask.java
@@ -1,4 +1,4 @@
-package Edith;
+package edith.task;
 
 public class ToDoTask extends Task {
     public ToDoTask(String taskName) {
